Refactor `fs_tools` structure: move tools and dependencies into a dedicated submodule and update tests accordingly
